+++ src/components/icon/assets/arrow_right.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-const EuiIconArrowRight = ({ title, ...props }) => (
+const EuiIconArrowRight = ({ title, titleId, ...props }) => (
   <svg
     width={16}
     height={16}
     viewBox="0 0 16 16"
     xmlns="http://www.w3.org/2000/svg"
+    aria-labelledby={titleId}
     {...props}>
-    <title>{title}</title>
+    {title ? <title id={titleId}>{title}</title> : null}
     <path
       fillRule="nonzero"
       d="M5.157 13.069l4.611-4.685a.546.546 0 000-.768L5.158 2.93a.552.552 0 010-.771.53.53 0 01.759 0l4.61 4.684c.631.641.63 1.672 0 2.312l-4.61 4.684a.53.53 0 01-.76 0 .552.552 0 010-.771z"