+++ src/components/icon/assets/editor_align_center.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-const EuiIconEditorAlignCenter = ({ title, ...props }) => (
+const EuiIconEditorAlignCenter = ({ title, titleId, ...props }) => (
   <svg
     width={16}
     height={16}
     viewBox="0 0 16 16"
     xmlns="http://www.w3.org/2000/svg"
+    aria-labelledby={titleId}
     {...props}>
-    <title>{title}</title>
+    {title ? <title id={titleId}>{title}</title> : null}
     <path d="M5 4V3h6v1H5zM3 7V6h10v1H3zm2 3V9h6v1H5zm-2 3v-1h10v1H3z" />
   </svg>
 );