+++ src/components/icon/assets/logo_uptime.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-const EuiIconLogoUptime = ({ title, ...props }) => (
+const EuiIconLogoUptime = ({ title, titleId, ...props }) => (
   <svg
     width={32}
     height={32}
     viewBox="0 0 32 32"
     xmlns="http://www.w3.org/2000/svg"
+    aria-labelledby={titleId}
     {...props}>
-    <title>{title}</title>
+    {title ? <title id={titleId}>{title}</title> : null}
     <path
       fill="#3EBEB0"
       d="M19 15c0 7.062-4.888 12.969-11.46 14.563A15.914 15.914 0 0016 32c8.836 0 16-7.163 16-16v-3l-6.5-6-6.5 6v2z"