+++ src/components/icon/assets/ml_create_advanced_job.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-const EuiIconMlCreateAdvancedJob = ({ title, ...props }) => (
+const EuiIconMlCreateAdvancedJob = ({ title, titleId, ...props }) => (
   <svg
     width={32}
     height={32}
     viewBox="0 0 32 32"
     xmlns="http://www.w3.org/2000/svg"
+    aria-labelledby={titleId}
     {...props}>
-    <title>{title}</title>
+    {title ? <title id={titleId}>{title}</title> : null}
     <path d="M16 32C7.163 32 0 24.837 0 16S7.163 0 16 0s16 7.163 16 16h-2c0-7.732-6.268-14-14-14S2 8.268 2 16s6.268 14 14 14v2z" />
     <path
       className="euiIcon__fillSecondary"