+++ src/components/icon/assets/tableOfContents.js
@@ -1,13 +1,14 @@
 import React from 'react';
 
-const EuiIconTableOfContents = ({ title, ...props }) => (
+const EuiIconTableOfContents = ({ title, titleId, ...props }) => (
   <svg
     width={16}
     height={16}
     viewBox="0 0 16 16"
     xmlns="http://www.w3.org/2000/svg"
+    aria-labelledby={titleId}
     {...props}>
-    <title>{title}</title>
+    {title ? <title id={titleId}>{title}</title> : null}
     <path d="M1 1v14h14V1H1zM0 0h16v16H0V0zm9 1v14h1V1H9zM3 3.5h4v-1H3v1zm0 3h4v-1H3v1zm0 3h4v-1H3v1z" />
   </svg>
 );